VS-44 Take out the ThemeSwitcher component
